# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: RFID-BACK API
  version: 1.0.0
  description: API de gestion des cueilleurs, tags, activités et conformités
servers:
  - url: http://localhost:8080/api
    description: Serveur local de développement
  - url: https://api.apolog.fr/api
    description: Serveur de production

# Default: a logged-in user session. /tags/scan uses the reader token, /auth/login is public.
security:
  - SessionCookie: []

# -----------------------------------------------------------------------------------------------
# Endpoints section
# -----------------------------------------------------------------------------------------------
paths:
...
  /tags/buckets/{bucketNumber}:
    post:
      summary: Register tags for a bucket
      description: >
        Add the provided tags to the bucket (created if unknown). Tags already linked to this
        bucket stay linked. Tags linked to another bucket are moved only when moveConfirmed is true;
        otherwise the request is refused with 409 and nothing is saved.
      operationId: registerTagsForBucket
      tags: [Tag]
      parameters:
        - in: path
          name: bucketNumber
          required: true
          description: Bucket number to associate with the provided tags
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterTagsRequest"
      responses:
        "200":
          description: Tags registered for the bucket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterTagsResponse"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Some tags belong to another bucket and the move was not confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagsInOtherBuckets"
        "500": { $ref: "#/components/responses/InternalError" }
  /tags/scan:
    post:
      summary: Scan a tag and report its compliance state
      description: >
        Endpoint called by an RFID reader to report whether a scanned tag is compliant.
        Accessible only with the API token generated during reader registration.
        401 is also returned when the reader is disabled.
        For a reader in ENREGISTREMENT mode, the read is kept for tag registration (or ignored when
        no registration session is open), no record is created, and isCompliant is always true.
      operationId: scanTag
      tags: [Tag]
      security:
        - ReaderApiToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanTagRequest"
      responses:
        "200":
          description: |
            Tag scanned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanTagResponse"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }

  # -----------------------------------------------------------------------------------------------
  # Registration section (tag registration sessions, spec 003)
  # -----------------------------------------------------------------------------------------------
  /tags/registration-sessions:
    post:
      summary: Start a tag registration session on a reader
      description: >
        The reader must be active and in ENREGISTREMENT mode. From now on, its scans are kept as
        temporary reads of this session. One session per reader; it expires after a period
        without activity.
      operationId: startRegistrationSession
      tags: [Registration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRegistrationSession"
      responses:
        "201":
          description: Session started, with no read yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationSession"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The reader already has an open session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReaderBusy"
  /tags/registration-sessions/{sessionId}:
    parameters:
      - in: path
        name: sessionId
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a registration session and the tags read since Start
      description: Only the Administrateur who started the session can read it.
      operationId: getRegistrationSession
      tags: [Registration]
      responses:
        "200":
          description: Session with its reads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationSession"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: Cancel a registration session and discard its reads
      operationId: cancelRegistrationSession
      tags: [Registration]
      responses:
        "204": { $ref: "#/components/responses/Deleted" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /tags/registration-sessions/{sessionId}/save:
    parameters:
      - in: path
        name: sessionId
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Register the session's tags on a bucket and close the session
      description: >
        Same rules as POST /tags/buckets/{bucketNumber}. On 409 nothing is saved and the session
        stays open.
      operationId: saveRegistrationSession
      tags: [Registration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveRegistrationSession"
      responses:
        "200":
          description: Tags registered; the session is closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterTagsResponse"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Some tags belong to another bucket and the move was not confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagsInOtherBuckets"

  # -----------------------------------------------------------------------------------------------
  # Record section
  # -----------------------------------------------------------------------------------------------
...
components:
...
  responses:
...
    Deleted:
      description: Retourné lorsqu'une ressource a été supprimée
...
    NotFound:
      description: Retourné lorsqu'une ressource est introuvable
    InvalidRequest:
      description: Requête invalide
    InternalError:
      description: Erreur interne côté serveur
...
    Unauthorized:
      description: Jeton API manquant ou invalide, ou utilisateur non connecté
    Forbidden:
      description: Rôle insuffisant

  # ---------------------------------------------------------------------------
...
  schemas:
...
    ScanTagRequest:
      type: object
      properties:
        uid:
          type: string
          description: Unique identifier of the scanned tag
          example: E2000017221101891400A23G
        isCompliant:
          type: boolean
          description: Tells whether the scanned tag is compliant
          example: false
      required:
        - uid
        - isCompliant
    ScanTagResponse:
      type: object
      properties:
        uid:
          type: string
          description: Unique identifier of the scanned tag
          example: E2000017221101891400A23G
        isCompliant:
          type: boolean
          description: Tells whether the scanned tag is compliant
          example: false
        processedAt:
          type: string
          format: date-time
          description: Scan processing date and time
          example: "2024-01-15T09:30:00Z"
        message:
          type: string
          description: Compliance processing message
          example: Tag marked as non-compliant due to missing validation
      required:
        - uid
        - isCompliant
        - processedAt
    RegisterTagsRequest:
      type: object
      properties:
        uids:
          type: array
          description: Unique IDs of scanned tags. At most 100 per registration; more is refused with 400.
          minItems: 1
          maxItems: 100
          items:
            type: string
            minLength: 1
            example: E2000017221101891400A23G
        moveConfirmed:
          type: boolean
          default: false
          description: >
            Must be true to move tags that currently belong to another bucket.
            Otherwise the request is refused with 409 and nothing is saved.
      required:
        - uids
    RegisterTagsResponse:
      type: object
      properties:
        bucketNumber:
          type: integer
          description: Bucket number associated with the registered tags
          example: 12
        registeredCount:
          type: integer
          description: Number of unique, non-blank tag uids in this request
          example: 4
        totalCount:
          type: integer
          description: Number of tags linked to the bucket after this registration
          example: 6
      required:
        - bucketNumber
        - registeredCount
        - totalCount
    TagInOtherBucket:
      type: object
      properties:
        uid:
          type: string
          example: E2000017221101891400A23G
        bucketNumber:
          type: integer
          example: 7
      required:
        - uid
        - bucketNumber
    TagsInOtherBuckets:
      type: object
      description: 409 body when tags would be moved from another bucket without confirmation
      properties:
        message:
          type: string
        tags:
          type: array
          items:
            $ref: "#/components/schemas/TagInOtherBucket"
      required:
        - message
        - tags
    StartRegistrationSession:
      type: object
      properties:
        readerId:
          type: string
          format: uuid
      required:
        - readerId
    RegistrationRead:
      type: object
      properties:
        uid:
          type: string
          example: E2000017221101891400A23G
        firstReadAt:
          type: string
          format: date-time
        bucketNumber:
          type: integer
          nullable: true
          description: Bucket the tag currently belongs to, or null if none
      required:
        - uid
        - firstReadAt
    RegistrationSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        readerId:
          type: string
          format: uuid
        readerUid:
          type: string
        startedBy:
          type: string
          description: Username of the Administrateur who started the session
        startedAt:
          type: string
          format: date-time
        reads:
          type: array
          description: Tags read since Start, one entry per tag, oldest first
          items:
            $ref: "#/components/schemas/RegistrationRead"
      required:
        - id
        - readerId
        - readerUid
        - startedBy
        - startedAt
        - reads
    ReaderBusy:
      type: object
      description: >
        409 body when the reader already has an open registration session. startedBy and startedAt are
        absent in the rare case where the reader stayed contended but its session could not be read back.
      properties:
        message:
          type: string
        startedBy:
          type: string
        startedAt:
          type: string
          format: date-time
      required:
        - message
    SaveRegistrationSession:
      type: object
      properties:
        bucketNumber:
          type: integer
          minimum: 1
        moveConfirmed:
          type: boolean
          default: false
      required:
        - bucketNumber
...
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Session opened by POST /auth/login. State-changing requests also need the X-XSRF-TOKEN header (value of the XSRF-TOKEN cookie).
    ReaderApiToken:
      type: apiKey
      in: header
      name: x-api-token
      description: API token assigned to the reader during its registration